Migration to Bootstrap 4 was required to get Webpacker happy.
This is a big change that I was hoping to avoid, because now we are on Bootstrap 4 for the /admin screens, but Bootstrap 3 for the main app.   They *appear* to be working together, since 3 is loaded via a Rails Gem and 4 is loaded via NPM.   However, our global CSS didn't migrate perfectly, so the navbar and screen layouts on the /admin pages are mucked up.   Tests do all run!