WIP: add whitelist for search hosts
This is a WIP, for early feedback.

Add a whitelist configuration to Quepid that restricts which hosts are
allowed to be used as search backends.

TODO: use the list in the JS code as well, to provide user guidance and
ensure they know early on what the list of valid hosts are.