Refactoring a bit...
The customize_quepid class is where we expect to find all customozations possible, so I moved the check to there.   I wanted a more generic handling of emails, so rename smtp_mail_provider to email_provider, That led to moving some more of the email settings stuff there.    Tweak the docs!