Added functionality to send mails over SMTP and start with an operating documentation (#275) (#276)
* Added functionality to send mails over smtp and start with an operating documentation

* Refactoring a bit...

The customize_quepid class is where we expect to find all customozations possible, so I moved the check to there.   I wanted a more generic handling of emails, so rename smtp_mail_provider to email_provider, That led to moving some more of the email settings stuff there.    Tweak the docs!

Co-authored-by: epugh@opensourceconnections.com <>