Heavy refactoring of how we get cases back to solve the left outer joins not owrking.
Our outer joins were returning MANY MORE cases than was approprite, with lots of duplicates.   This was introduced when we removed the any_of gem as part of upgrading to Rails 5, and the left outer join wasn't done right.  Didn't matter with a few cases, but eric's account should return 141 cases and was returnign 550 cases, which made the API slow, causing Heroku to kill it.