Restore proper support for notifying users about Cookies being required policy (#1127)
The cookie logic no longer worked, and was only run on the cases page, which was counter intuitive.  The links to the cookie policy didn't render either.  Now we get a Toast notification ON the homepage, and if you accept the warning, you don't get prompted OVER AND OVER.  We are able to remove two gems in favour of a npm package and some simple JS.